# Fix type instability in `getproperty(::Schema, ::Symbol)`