@bkamins 2023-08-05
I have two comments:
1. Have you benchmarked how `Val` scales for wide tables (like 1000+ columns)?
2. Could you please add a test that checks for type stability of the result?