@MilesCranmer 2023-08-05
> 1. Have you benchmarked how Val scales for wide tables (like 1000+ columns)?

Keep in mind that this PR does not modify the code's behavior at all. It just switches from `Tuple(... for i=1:n)` to `ntuple(..., Val{n})`, taking advantage of the fact that `n` is a compile-time constant (if `types` is not `nothing`). The return type is the same, it just helps the compiler infer it.

If `types` are nothing, as I guess would happen for a large table like that (?), the `Val` would not even get hit:
```julia
return types === nothing ? (T !== nothing ? T : nothing) : ntuple(i -> fieldtype(types, i), Val(fieldcount(types)))
```

> 2. Could you please add a test that checks for type stability of the result?

Any tips on how to do this? I've never unit-tested compiler type inference before. It feels like something that would be closely tied to a specific Julia version too.