@MilesCranmer 2023-08-05
Oh it looks like there is `Test.@inferred`. Is that what you mean?