@bkamins 2023-08-05
> The return type is the same, it just helps the compiler infer it.

The point of my question is that I want to make sure that if the compiler tries to infer it it is not too costly. (maybe it does not make a difference, but I want to make sure; in general passing `Val` to `ntuple` usually leads to code generation).

> Oh it looks like there is `Test.@inferred`. Is that what you mean?

Yes