@MilesCranmer 2023-08-05
> The point of my question is that I want to make sure that if the compiler tries to infer it it is not too costly. (maybe it does not make a difference, but I want to make sure; in general passing `Val` to `ntuple` usually leads to code generation).

Okay got it. I just tried and there's no difference. I used the `WideTable` unittest to check. Here is the before:
```
julia> @benchmark getproperty(s, :types) setup=(w=WideTable(); s=Tables.schema(w))
BenchmarkTools.Trial: 177 samples with 995 evaluations.
 Range (min … max):  39.112 ns … 42.797 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     39.866 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   39.933 ns ±  0.599 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅▁        ▂▁ █  ▁▆                                           
  ██▆▁▃▄▄▆▅▄██▇█████▆▆█▄▄▄▄▃▄▃▃▄▃▁▃▄▃▃▁▁▄▁▃▁▁▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▃ ▃
  39.1 ns         Histogram: frequency by time        42.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
and the after
```
julia> @benchmark getproperty(s, :types) setup=(w=WideTable(); s=Tables.schema(w))
BenchmarkTools.Trial: 180 samples with 997 evaluations.
 Range (min … max):  38.406 ns … 60.265 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     39.222 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   39.552 ns ±  1.945 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▁   ▆█▇▆▂                                                    
  █▃▃▆█████▆▆▆▃▁▃▃▃▃▃▁▁▃▁▁▁▁▁▃▃▃▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▃
  38.4 ns         Histogram: frequency by time        45.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```


> Yes

Done!