@MilesCranmer 2023-08-05
For good measure, here's a normal table with 100 columns:

Before:
```
julia> @benchmark getproperty(s, :types) setup=(x=NamedTuple([gensym() => [1.0] for _=1:100]); s=Tables.schema(x))
BenchmarkTools.Trial: 317 samples with 1 evaluation.
 Range (min … max):  8.034 ms … 28.657 ms  ┊ GC (min … max): 0.00% … 68.35%
 Time  (median):     8.170 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.299 ms ±  1.380 ms  ┊ GC (mean ± σ):  1.22% ±  5.03%

     ▄▂▂▆██▆▂▄▆                                               
  ▆▃████████████▆▇▅▆▅▆▃▃▂▂▂▃▂▁▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▃▁▁▁▁▁▁▁▁▂ ▄
  8.03 ms        Histogram: frequency by time         8.9 ms <

 Memory estimate: 1.18 MiB, allocs estimate: 17500.
```

and after:

```
julia> @benchmark getproperty(s, :types) setup=(x=NamedTuple([gensym() => [1.0] for _=1:100]); s=Tables.schema(x))
BenchmarkTools.Trial: 323 samples with 1 evaluation.
 Range (min … max):  7.970 ms …  21.561 ms  ┊ GC (min … max): 0.00% … 59.28%
 Time  (median):     8.070 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.145 ms ± 766.230 μs  ┊ GC (mean ± σ):  0.49% ±  3.30%

     ▂█▅▄▂                                                     
  ▃▄▇█████▆▆▅▃▄▄▄▄▂▃▃▃▁▂▁▁▂▁▃▁▁▁▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▃
  7.97 ms         Histogram: frequency by time        8.95 ms <

 Memory estimate: 135.90 KiB, allocs estimate: 1842.
```

However, if I go all the way to 1000 columns, then it does starts to hurt performance, as you guessed.

Here's the before:

```
julia> @benchmark getproperty(s, :types) setup=(x=NamedTuple([gensym() => [1.0] for _=1:1000]); s=Tables.schema(x))
BenchmarkTools.Trial: 254 samples with 1 evaluation.
 Range (min … max):  8.088 ms … 22.118 ms  ┊ GC (min … max): 0.00% … 55.82%
 Time  (median):     8.327 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.665 ms ±  1.088 ms  ┊ GC (mean ± σ):  0.87% ±  4.43%

   █▅▃▃▆▁                                                     
  ███████▅▅▇▆▄▃▁▆▃▃▄▄▄▃▄▃▃▃▄▄▃▃▃▃▄▃▄▃▃▁▃▅▃▃▅▄▃▃▃▃▃▃▁▃▃▁▃▁▁▃▃ ▃
  8.09 ms        Histogram: frequency by time          10 ms <

 Memory estimate: 1.20 MiB, allocs estimate: 17992.
```

and the after

```
julia> @benchmark getproperty(s, :types) setup=(x=NamedTuple([gensym() => [1.0] for _=1:1000]); s=Tables.schema(x))
BenchmarkTools.Trial: 28 samples with 1 evaluation.
 Range (min … max):  169.831 ms … 177.141 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     171.580 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   171.891 ms ±   1.547 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

        █ ▃   ▃▃ ▃  ▃   ▃              ▃                         
  ▇▁▁▇▁▇█▇█▁▇▁██▁█▇▇█▁▁▁█▇▁▁▇▁▁▁▁▁▇▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇ ▁
  170 ms           Histogram: frequency by time          177 ms <

 Memory estimate: 175.15 KiB, allocs estimate: 1843.
```

Thoughts?