@MilesCranmer 2023-08-05
Oh, wait, I think that accidentally included compilation time. If you fix the table, it's just fine. It even improves performance:

Before:
```
julia> x = NamedTuple([gensym() => [1.0] for _=1:1000]);

julia> s = Tables.schema(x);

julia> @benchmark getproperty(s, :types)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  15.708 μs … 249.708 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     17.208 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   17.460 μs ±   3.971 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

         ▁█▆▂                                                   
  ▃▆▄▃▃▂▄████▇▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▁▂▂▂▂ ▃
  15.7 μs         Histogram: frequency by time         25.7 μs <

 Memory estimate: 15.81 KiB, allocs estimate: 2.
```

After:
```
julia> x = NamedTuple([gensym() => [1.0] for _=1:1000]);

julia> s = Tables.schema(x);

julia> @benchmark getproperty(s, :types)
BenchmarkTools.Trial: 10000 samples with 225 evaluations.
 Range (min … max):  343.333 ns … 651.208 μs  ┊ GC (min … max):  0.00% … 99.74%
 Time  (median):       1.053 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):     1.776 μs ±  14.498 μs  ┊ GC (mean ± σ):  35.42% ±  4.56%

                     ▃▄▇▇██▇▆▅▄▃▂▁  ▂▃▃▃▃▃▃▃▂▂▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁   ▃
  ▆▅▁▄▃▁▁▃▄▅▁▅▁▄▄▄▅▆▇█████████████▇████████████████████████████ █
  343 ns        Histogram: log(frequency) by time       2.12 μs <

 Memory estimate: 7.88 KiB, allocs estimate: 1.
```